requirements.txt
openai~=0.27.2
python-decouple~=3.8
python-telegram-bot~=13.0
